# SourceCode/front-end/logsmart/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "back-end", "description": "", "license": { "name": "" }, "version": "1.1.0" },
	"paths": {
...
		"/logs/admin/entries": {
			"get": {
				"tags": ["Log Entries"],
				"summary": "Lists all log entries for the company or branch (managers only).",
				"operationId": "list_company_log_entries",
				"responses": {
					"200": {
						"description": "Company log entries retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ListLogEntriesResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - only managers can view all entries",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/entries": {
			"get": {
				"tags": ["Log Entries"],
				"summary": "Lists all log entries for the current user, optionally filtered by template or status.",
				"description": "# Errors\nReturns an error if the user is not authorized or if the query fails.",
				"operationId": "list_user_log_entries",
				"responses": {
					"200": {
						"description": "User log entries retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ListLogEntriesResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			},
			"post": {
				"tags": ["Log Entries"],
				"summary": "Creates a new log entry draft.",
				"description": "# Errors\nReturns an error if the template name is empty or if entry creation fails.",
				"operationId": "create_log_entry",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/CreateLogEntryRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Log entry created successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/CreateLogEntryResponse" }
							}
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Template not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/entries/due": {
			"get": {
				"tags": ["Log Entries"],
				"summary": "Lists all log forms that are due today for the current user's company.",
				"description": "# Errors\nReturns an error if the user is not authorized or if the query fails.",
				"operationId": "list_due_forms_today",
				"responses": {
					"200": {
						"description": "Due forms retrieved successfully",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/DueFormsResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/entries/{entry_id}": {
			"get": {
				"tags": ["Log Entries"],
				"summary": "Retrieves a specific log entry by its ID.",
				"description": "# Errors\nReturns an error if the entry is not found or if the user is not authorized.",
				"operationId": "get_log_entry",
				"parameters": [
					{ "name": "entry_id", "in": "path", "required": true, "schema": { "type": "string" } }
				],
				"responses": {
					"200": {
						"description": "Log entry retrieved successfully",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/LogEntryResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - entry does not belong to user",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Entry not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			},
			"put": {
				"tags": ["Log Entries"],
				"summary": "Updates the data of an existing log entry draft.",
				"description": "# Errors\nReturns an error if the entry is not found or if the update fails.",
				"operationId": "update_log_entry",
				"parameters": [
					{ "name": "entry_id", "in": "path", "required": true, "schema": { "type": "string" } }
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/UpdateLogEntryRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Log entry updated successfully",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/LogEntryResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - entry does not belong to user",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Entry not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			},
			"delete": {
				"tags": ["Log Entries"],
				"summary": "Deletes a log entry.",
				"description": "# Errors\nReturns an error if the user is not authorized or if deletion fails.",
				"operationId": "delete_log_entry",
				"parameters": [
					{ "name": "entry_id", "in": "path", "required": true, "schema": { "type": "string" } }
				],
				"responses": {
					"200": {
						"description": "Log entry deleted successfully",
						"content": { "application/json": { "schema": {} } }
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - entry does not belong to user",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Entry not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/entries/{entry_id}/submit": {
			"post": {
				"tags": ["Log Entries"],
				"summary": "Submits a log entry, marking it as final.",
				"description": "# Errors\nReturns an error if the entry is not found or if submission fails.",
				"operationId": "submit_log_entry",
				"parameters": [
					{ "name": "entry_id", "in": "path", "required": true, "schema": { "type": "string" } }
				],
				"responses": {
					"200": {
						"description": "Log entry submitted successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/SubmitLogEntryResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - entry does not belong to user",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Entry not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/logs/entries/{entry_id}/unsubmit": {
			"post": {
				"tags": ["Log Entries"],
				"summary": "Returns a submitted log entry to draft status (admin only).",
				"description": "# Errors\nReturns an error if the entry is not found or if the operation fails.",
				"operationId": "unsubmit_log_entry",
				"parameters": [
					{ "name": "entry_id", "in": "path", "required": true, "schema": { "type": "string" } }
				],
				"responses": {
					"200": {
						"description": "Log entry returned to draft successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/SubmitLogEntryResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - only admins can unsubmit entries",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Entry not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CreateLogEntryRequest": {
				"type": "object",
				"required": ["template_name"],
				"properties": { "template_name": { "type": "string", "example": "Kitchen Daily Log" } }
			},
			"CreateLogEntryResponse": {
				"type": "object",
				"required": ["id", "message"],
				"properties": { "id": { "type": "string" }, "message": { "type": "string" } }
			},
...
			"DueFormsResponse": {
				"type": "object",
				"required": ["forms"],
				"properties": {
					"forms": { "type": "array", "items": { "$ref": "#/components/schemas/DueFormInfo" } }
				}
			},
			"ErrorResponse": {
				"type": "object",
				"required": ["error"],
				"properties": { "error": { "type": "string" } }
			},
...
			"ListLogEntriesResponse": {
				"type": "object",
				"required": ["entries"],
				"properties": {
					"entries": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/LogEntryResponse" }
					}
				}
			},
			"LogEntryResponse": {
				"type": "object",
				"required": [
					"id",
					"template_name",
					"template_layout",
					"entry_data",
					"status",
					"created_at",
					"updated_at",
					"period"
				],
				"properties": {
					"created_at": { "type": "string" },
					"entry_data": {},
					"id": { "type": "string" },
					"period": { "type": "string" },
					"status": { "type": "string" },
					"submitted_at": { "type": ["string", "null"] },
					"template_layout": { "$ref": "#/components/schemas/Vec" },
					"template_name": { "type": "string" },
					"updated_at": { "type": "string" }
				}
			},
...
			"SubmitLogEntryResponse": {
				"type": "object",
				"required": ["message"],
				"properties": { "message": { "type": "string" } }
			},
...
			"UpdateLogEntryRequest": {
				"type": "object",
				"required": ["entry_data"],
				"properties": { "entry_data": {} }
			},
...
		},
		"securitySchemes": { "bearer_auth": { "type": "http", "scheme": "bearer" } }
	},
...
}